Extract newsletter.js
Instead of using an inline script, we move the JavaScript into its own
file.  Since the script uses some language strings, we put them into a
data- attribute of the script element, so that we can easily access the
information from JavaScript.  We also do that to signal whether the
subscriber fields are mandatory.

Note that so far we do not cater to the many deficiencies of the
existing JS (most notably that the script introduces several global
variables which are not even "namespaced"), nor do we cater to
efficiency (the script should either be written to `$bjs` or defered).

Note also that the script now requires a somewhat contemporary browser,
while previously it was supposed to work even in ES3 environments (e.g.
IE 6).  If that is a concern, the script could be adapted.